把註冊元件變成 component
